@@ -1,9 +1,9 @@
 name: graphic_counter
-description: Counting application displaying historical data as a graph 
+description: Counting application displaying historical data as a graph
 
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
-publish_to: 'none' # Remove this line if you wish to publish to pub.dev
+publish_to: "none" # Remove this line if you wish to publish to pub.dev
 
 # The following defines the version and build number for your application.
 # A version number is three numbers separated by dots, like 1.2.43
@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.1.2+7
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -30,7 +30,6 @@ dependencies:
   flutter:
     sdk: flutter
 
-
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
@@ -38,7 +37,7 @@ dependencies:
   firebase_core: ^3.13.0
   flutter_palette: ^1.1.0+1
   adaptive_dialog: ^2.4.1
-  firebase_auth: ^5.5.2
+  firebase_auth: ^5.5.3
   google_sign_in: ^6.3.0
   cloud_firestore: ^5.6.6
   path_provider: ^2.0.7
@@ -49,7 +48,12 @@ dependencies:
   download: ^1.0.1
   flutter_launcher_icons: ^0.14.3
   color_palette_plus: ^2.0.0
-  upgrade: ^0.0.8
+  google_sign_in_platform_interface: ^2.5.0
+  google_sign_in_web: ^0.12.4+4
+  firebase_ui_auth: ^1.16.1
+  firebase_ui_oauth_google: ^1.4.1
+  google_fonts: ^6.2.1
+  collection: ^1.19.1
 
 dev_dependencies:
   flutter_test:
